refactor(Package): Moved fetch license out of modal to populate license fullname at edit page
